Move GoogleDiscoveryApiHook to google.common
