+++ tests/providers/google/cloud/hooks/test_discovery_api.py
@@ -21,7 +21,7 @@ from unittest.mock import call, patch
 
 from airflow import models
 from airflow.configuration import load_test_config
-from airflow.providers.google.cloud.hooks.discovery_api import GoogleDiscoveryApiHook
+from airflow.providers.google.common.hooks.discovery_api import GoogleDiscoveryApiHook
 from airflow.utils import db
 
 
@@ -40,8 +40,8 @@ class TestGoogleDiscoveryApiHook(unittest.TestCase):
             )
         )
 
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.build')
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.GoogleDiscoveryApiHook._authorize')
+    @patch('airflow.providers.google.common.hooks.discovery_api.build')
+    @patch('airflow.providers.google.common.hooks.discovery_api.GoogleDiscoveryApiHook._authorize')
     def test_get_conn(self, mock_authorize, mock_build):
         google_discovery_api_hook = GoogleDiscoveryApiHook(
             gcp_conn_id='google_test',
@@ -58,8 +58,8 @@ class TestGoogleDiscoveryApiHook(unittest.TestCase):
             cache_discovery=False
         )
 
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.getattr')
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.GoogleDiscoveryApiHook.get_conn')
+    @patch('airflow.providers.google.common.hooks.discovery_api.getattr')
+    @patch('airflow.providers.google.common.hooks.discovery_api.GoogleDiscoveryApiHook.get_conn')
     def test_query(self, mock_get_conn, mock_getattr):
         google_discovery_api_hook = GoogleDiscoveryApiHook(
             gcp_conn_id='google_test',
@@ -90,8 +90,8 @@ class TestGoogleDiscoveryApiHook(unittest.TestCase):
             call()().execute(num_retries=num_retries)
         ])
 
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.getattr')
-    @patch('airflow.providers.google.cloud.hooks.discovery_api.GoogleDiscoveryApiHook.get_conn')
+    @patch('airflow.providers.google.common.hooks.discovery_api.getattr')
+    @patch('airflow.providers.google.common.hooks.discovery_api.GoogleDiscoveryApiHook.get_conn')
     def test_query_with_pagination(self, mock_get_conn, mock_getattr):
         google_api_conn_client_sub_call = mock_getattr.return_value.return_value
         mock_getattr.return_value.side_effect = [
